+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- new page to search and select merged and extracted items
 - highlight required fields in edit view with a red asterisk
 - add support for logging-in with ENTER key